use pin macros
- misc changes
- remove object_get_header usage

+++ xlive/Blam/Engine/main/main_game.cpp
@@ -187,7 +187,7 @@ void main_game_launch_setup_game_mode_details(void)
 #ifdef MSVC
         error(2, __FUNCTION__": unknown game mode %d!", g_main_game_launch_options.game_mode);
 #else
-        error(2, "%s: unknown game mode %d!", __FUNCTION__, g_main_game_launch_options.game_mode);
+        error(2, "%s: unknown game mode %d!", __FUNCTION__, (uint32)g_main_game_launch_options.game_mode);
 #endif
     }
     }
